Rename webhook adapters to drivers, split multi-export types, drop mocked service specs
