+++ packages/twenty-server/src/modules/messaging-webhooks/drivers/aws-ses/services/ses-inbound-webhook-driver.service.ts
@@ -3,23 +3,21 @@ import { Injectable, Logger } from '@nestjs/common';
 import type SnsPayloadValidator from 'sns-payload-validator';
 import { isDefined, parseJson } from 'twenty-shared/utils';
 
-import { SnsSignatureVerifierService } from 'src/modules/messaging-webhooks/adapters/aws-ses/services/sns-signature-verifier.service';
-import { SnsSubscriptionConfirmerService } from 'src/modules/messaging-webhooks/adapters/aws-ses/services/sns-subscription-confirmer.service';
-import { type SesInboundNotification } from 'src/modules/messaging-webhooks/adapters/aws-ses/types/sns-message.type';
+import { SnsSignatureVerifierService } from 'src/modules/messaging-webhooks/drivers/aws-ses/services/sns-signature-verifier.service';
+import { SnsSubscriptionConfirmerService } from 'src/modules/messaging-webhooks/drivers/aws-ses/services/sns-subscription-confirmer.service';
+import { type SesInboundNotification } from 'src/modules/messaging-webhooks/drivers/aws-ses/types/sns-message.type';
 import { InboundMailHandlerService } from 'src/modules/messaging-webhooks/handlers/inbound-mail-handler.service';
 import { MessagingWebhookExceptionCode } from 'src/modules/messaging-webhooks/messaging-webhook-exception-code.enum';
 import { MessagingWebhookException } from 'src/modules/messaging-webhooks/messaging-webhook.exception';
 import { type NormalizedInboundMailNotification } from 'src/modules/messaging-webhooks/types/normalized-inbound-mail-notification.type';
-import {
-  type InboundEmailMessageReference,
-  INBOUND_EMAIL_MESSAGE_SOURCE,
-} from 'src/modules/messaging/message-import-manager/drivers/inbound-email/types/inbound-email-message-source.type';
+import { INBOUND_EMAIL_MESSAGE_SOURCE } from 'src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email-message-source.constant';
+import { type InboundEmailMessageReference } from 'src/modules/messaging/message-import-manager/drivers/inbound-email/types/inbound-email-message-reference.type';
 
 type SnsPayload = SnsPayloadValidator.SnsPayload;
 
 @Injectable()
-export class SesInboundWebhookAdapterService {
-  private readonly logger = new Logger(SesInboundWebhookAdapterService.name);
+export class SesInboundWebhookDriverService {
+  private readonly logger = new Logger(SesInboundWebhookDriverService.name);
 
   constructor(
     private readonly snsSignatureVerifierService: SnsSignatureVerifierService,
@@ -71,25 +69,22 @@ export class SesInboundWebhookAdapterService {
     snsMessageId: string,
   ): NormalizedInboundMailNotification {
     const action = notification.receipt?.action;
+    const recipients = notification.receipt?.recipients ?? [];
+    const subject = notification.mail?.commonHeaders?.subject ?? null;
 
-    let message: InboundEmailMessageReference | null = null;
-
-    if (action?.type === 'S3') {
-      message = {
-        source: INBOUND_EMAIL_MESSAGE_SOURCE.SES_S3,
-        reference: action.objectKey,
-      };
-    } else {
+    if (action?.type !== 'S3') {
       this.logger.warn(
         `SNS message ${snsMessageId} has unsupported action type ${action?.type}`,
       );
+
+      return { recipients, subject, message: null, dedupeKey: snsMessageId };
     }
 
-    return {
-      recipients: notification.receipt?.recipients ?? [],
-      subject: notification.mail?.commonHeaders?.subject ?? null,
-      message,
-      dedupeKey: snsMessageId,
+    const message: InboundEmailMessageReference = {
+      source: INBOUND_EMAIL_MESSAGE_SOURCE.SES_S3,
+      reference: action.objectKey,
     };
+
+    return { recipients, subject, message, dedupeKey: snsMessageId };
   }
 }
