+++ packages/twenty-server/src/modules/messaging-webhooks/drivers/aws-ses/types/ses-event-bridge-notification.type.ts
