+++ packages/twenty-server/src/modules/messaging-webhooks/drivers/aws-ses/types/sns-message.type.ts
