+++ packages/twenty-server/src/modules/messaging-webhooks/types/normalized-outbound-sending-state-event.type.ts
@@ -1,0 +1,6 @@
+import { type EmailingDomainTenantStatus } from 'src/engine/core-modules/emailing-domain/drivers/types/emailing-domain-tenant-status.type';
+
+export type NormalizedOutboundSendingStateEvent = {
+  workspaceId: string;
+  status: EmailingDomainTenantStatus;
+};
