+++ packages/twenty-server/src/modules/messaging-webhooks/types/normalized-outbound-suppression-event.type.ts
@@ -1,4 +1,3 @@
-import { type EmailingDomainTenantStatus } from 'src/engine/core-modules/emailing-domain/drivers/types/emailing-domain-tenant-status.type';
 import { type MessageSuppressionReason } from 'src/engine/core-modules/emailing-domain/types/message-suppression-reason.type';
 
 export type NormalizedOutboundSuppressionEvent = {
@@ -7,9 +6,4 @@ export type NormalizedOutboundSuppressionEvent = {
   emailAddresses: string[];
   providerMessageId: string | null;
   providerEventId: string | null;
-};
-
-export type NormalizedOutboundSendingStateEvent = {
-  workspaceId: string;
-  status: EmailingDomainTenantStatus;
 };
