+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email-message-source.constant.ts
@@ -1,0 +1,3 @@
+export const INBOUND_EMAIL_MESSAGE_SOURCE = {
+  SES_S3: 'SES_S3',
+} as const;
