+++ packages/twenty-server/src/modules/messaging/message-import-manager/drivers/inbound-email/types/inbound-email-message-source.type.ts
@@ -1,11 +1,4 @@
-export const INBOUND_EMAIL_MESSAGE_SOURCE = {
-  SES_S3: 'SES_S3',
-} as const;
+import { type INBOUND_EMAIL_MESSAGE_SOURCE } from 'src/modules/messaging/message-import-manager/drivers/inbound-email/constants/inbound-email-message-source.constant';
 
 export type InboundEmailMessageSource =
   (typeof INBOUND_EMAIL_MESSAGE_SOURCE)[keyof typeof INBOUND_EMAIL_MESSAGE_SOURCE];
-
-export type InboundEmailMessageReference = {
-  source: InboundEmailMessageSource;
-  reference: string;
-};
